pyproject.toml
[project]
name = "resume-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "huggingface-hub>=0.34.0,<1.0",
    "python-docx>=1.2.0",
    "qdrant-client>=1.16.1",
]
